docs and move is spatial predicate supported to default join
